requirements.txt
chromadb
openai
python-dotenv
langchain-community
langchain-text-splitters
pypdf
langchain-openai
langchain-chroma
flask